Move BindGroupLayoutDescriptor, RenderPipelineDescriptor, NoFragmentStateError, VertexState, FragmentState, ComputePipelineDescriptor
