+++ crates/bevy_material/src/render_resource/mod.rs
@@ -1,10 +1,10 @@
 mod bind_group_layout_entries;
-// mod pipeline;
+mod pipeline;
 // mod pipeline_cache;
 mod pipeline_specializer;
 
 pub use bind_group_layout_entries::*;
-// pub use pipeline::*;
+pub use pipeline::*;
 // pub use pipeline_cache::*;
 pub use pipeline_specializer::*;
